Improve on CLI testability.
